requirements.txt
aiogram==3.0.0
python-dotenv==1.0.1
aiocryptopay==0.4.4
requests==2.32.3